requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cmdstanpy==1.2.4
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Django==5.1.2
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
fonttools==4.54.1
fpdf==1.7.2
holidays==0.59
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
kiwisolver==1.4.7
kombu==5.4.2
matplotlib==3.9.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
pluggy==1.5.0
prompt_toolkit==3.0.48
prophet==1.1.6
pyparsing==3.2.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
redis==5.1.1
reportlab==4.2.5
requests==2.32.3
ruff==0.7.1
six==1.16.0
sqlparse==0.5.1
stanio==0.5.1
tenacity==9.0.0
tqdm==4.66.5
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
